moved ct and wordpress to wordpress_v1 and wordpress_v2
