+++ lib/main.dart
@@ -1,13 +1,13 @@
-import 'package:app/services/ct/ct_ngg_gallery_service.dart';
-import 'package:app/services/ct/ct_rules_service.dart';
-import 'package:app/services/ct/ct_shortcode_service.dart';
-import 'package:app/services/ct/ct_sno_gallery_service.dart';
-import 'package:app/services/ct/ct_tab_category_service.dart';
+import 'package:app/services/wordpress/v1/ct_ngg_gallery_service.dart';
+import 'package:app/services/wordpress/v1/ct_rules_service.dart';
+import 'package:app/services/wordpress/v1/ct_shortcode_service.dart';
+import 'package:app/services/wordpress/v1/ct_sno_gallery_service.dart';
+import 'package:app/services/wordpress/v1/ct_tab_category_service.dart';
 import 'package:app/services/saved_posts_service.dart';
 import 'package:app/services/section/parser/shortcode_parser_service.dart';
-import 'package:app/services/wordpress/wordpress_media_service.dart';
-import 'package:app/services/wordpress/wordpress_posts_service.dart';
-import 'package:app/services/wordpress/wordpress_search_service.dart';
+import 'package:app/services/wordpress/v2/wordpress_media_service.dart';
+import 'package:app/services/wordpress/v2/wordpress_posts_service.dart';
+import 'package:app/services/wordpress/v2/wordpress_search_service.dart';
 import 'package:app/views/home_view.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
