+++ lib/services/section/sections/caption_section.dart
@@ -1,6 +1,6 @@
 import 'package:app/services/section/article_section.dart';
 import 'package:app/services/section/parser/shortcode_parser_service.dart';
-import 'package:app/services/wordpress/wordpress_media_service.dart';
+import 'package:app/services/wordpress/v2/wordpress_media_service.dart';
 import 'package:app/ui/transparent_image.dart';
 import 'package:flutter/material.dart';
