+++ lib/services/section/sections/pullquote_section.dart
@@ -1,6 +1,6 @@
 import 'package:app/services/section/article_section.dart';
 import 'package:app/services/section/parser/shortcode_parser_service.dart';
-import 'package:app/services/wordpress/wordpress_media_service.dart';
+import 'package:app/services/wordpress/v2/wordpress_media_service.dart';
 import 'package:flutter/material.dart';
 
 class PullquoteSection implements ArticleSection {
