+++ lib/ui/list/post_list/post_list_view.dart
@@ -1,5 +1,5 @@
 import 'package:app/models/post_model.dart';
-import 'package:app/services/wordpress/wordpress_posts_service.dart';
+import 'package:app/services/wordpress/v2/wordpress_posts_service.dart';
 import 'package:app/ui/list/list_page.dart';
 import 'package:app/ui/list/post_list/post_preview_card.dart';
 import 'package:flutter/material.dart';
