+++ lib/ui/list/post_list/post_preview_card.dart
@@ -1,8 +1,8 @@
 import 'dart:math';
 
 import 'package:app/models/post_model.dart';
-import 'package:app/services/wordpress/wordpress_media_service.dart';
-import 'package:app/ui/save_button.dart';
+import 'package:app/services/wordpress/v2/wordpress_media_service.dart';
+import 'package:app/ui/custom_buttons.dart';
 import 'package:app/views/article_view/article_view.dart';
 import 'package:app/views/featured_view.dart';
 import 'package:flutter/material.dart';
