+++ lib/ui/list/saved_list_view.dart
@@ -1,6 +1,6 @@
 import 'package:app/models/post_model.dart';
 import 'package:app/services/saved_posts_service.dart';
-import 'package:app/services/wordpress/wordpress_posts_service.dart';
+import 'package:app/services/wordpress/v2/wordpress_posts_service.dart';
 import 'package:app/ui/list/post_list/post_preview_card.dart';
 import 'package:app/ui/media_loading_indicator.dart';
 import 'package:flutter/material.dart';
